+++ pymqdatastream/doc/changelog.rst
@@ -1,5 +1,6 @@
 Changelog
 *********
-
+* 0.6.2: replaced json request/reply functionality with a ubjson dumpb/loadb version
+  responsible for all serialisation in _send_req_/_get_rep_
 * 0.6.11: First REQ/REP pair worked, added pymqds_bare_reqrep.py
 * 0.6.1: Added rudimentary support for REP/REQ stream functionality